pyproject.toml
[project]
name = "auction-shop"
version = "0.1.0"
description = ""
authors = [
    {name = "Viktoria Verkholat",email = "peremoga2009@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django (>=5.2.5,<6.0.0)",
    "ninja (>=1.13.0,<2.0.0)",
    "pillow (>=11.3.0,<12.0.0)",
    "django-widget-tweaks (>=1.5.0,<2.0.0)",
    "crispy (>=0.8.0,<0.9.0)",
    "pydantic (>=2.12.5,<3.0.0)",
    "factory-boy (>=3.3.3,<4.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
